:recycle:: move email rendering and SMTP delivery into domain and infrastructure layers
