+++ mcr-core/mcr_meeting/app/domain/email/__init__.py
@@ -1,0 +1,9 @@
+from mcr_meeting.app.domain.email.builder import (
+    build_report_ready_email,
+    build_transcription_ready_email,
+)
+
+__all__ = [
+    "build_report_ready_email",
+    "build_transcription_ready_email",
+]
