+++ mcr-core/mcr_meeting/app/domain/email/registry/footer.html
